+++ miden-crypto/src/dsa/falcon512_poseidon2/signature.rs
@@ -15,7 +15,7 @@ use crate::Word;
 // FALCON SIGNATURE
 // ================================================================================================
 
-/// A deterministic Poseidon2 Falcon512 signature over a message.
+/// A deterministic Falcon512 Poseidon2 signature over a message.
 ///
 /// The signature is a pair of polynomials (s1, s2) in (Z_p\[x\]/(phi))^2 a nonce `r`, and a public
 /// key polynomial `h` where:
@@ -38,7 +38,7 @@ use crate::Word;
 /// 1. the hash-to-point algorithm is made deterministic by using a fixed nonce `r`. This fixed
 ///    nonce is formed as `nonce_version_byte || preversioned_nonce` where `preversioned_nonce` is a
 ///    39-byte string that is defined as: i. a byte representing `log_2(512)`, followed by ii. the
-///    UTF8 representation of the string "Poseidon2-FALCON-DET", followed by iii. the required
+///    UTF8 representation of the string "FALCON-POSEIDON2-DET", followed by iii. the required
 ///    number of 0_u8 padding to make the total length equal 39 bytes. Note that the above means in
 ///    particular that only the `nonce_version_byte` needs to be serialized when serializing the
 ///    signature. This reduces the deterministic signature compared to the reference implementation
@@ -49,7 +49,7 @@ use crate::Word;
 /// The signature is serialized as:
 ///
 /// 1. A header byte specifying the algorithm used to encode the coefficients of the `s2` polynomial
-///    together with the degree of the irreducible polynomial phi. For Poseidon2 Falcon512, the
+///    together with the degree of the irreducible polynomial phi. For Falcon512 Poseidon2, the
 ///    header byte is set to `10111001` to differentiate it from the standardized instantiation of
 ///    the Falcon signature.
 /// 2. 1 byte for the nonce version.